refactor: WebSocket 대응 함수 이동
