+++ src/main.ts
@@ -3,10 +3,8 @@ import { join } from 'node:path';
 import serverlessExpress from '@codegenie/serverless-express';
 import { ClassSerializerInterceptor, Logger, ValidationPipe } from '@nestjs/common';
 import { HttpAdapterHost, NestFactory, Reflector } from '@nestjs/core';
-import { NestExpressApplication } from '@nestjs/platform-express';
 import { WsAdapter } from '@nestjs/platform-ws';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { APIGatewayEvent, Handler } from 'aws-lambda';
 
 import { AppModule } from './app';
 import { HttpException } from './common/exceptions';
@@ -15,7 +13,11 @@ import { MutationLogInterceptor, StringifyDateInterceptor, TimeoutInterceptor } 
 import { TypedConfigService } from './config/env';
 import { LoggerService } from './customs/logger';
 import { JwtAuthTokenService } from './modules/auth';
+import { configureWebSocketEvent } from './modules/event';
 import { PrismaService } from './prisma';
+
+import type { NestExpressApplication } from '@nestjs/platform-express';
+import type { Handler } from 'aws-lambda';
 
 export async function configure(app: NestExpressApplication) {
   const logger = await app.resolve(LoggerService);
@@ -119,24 +121,6 @@ if (process.env.SERVERLESS) {
   bootstrap = startServerless();
 } else if (process.env.STAGE === 'local') {
   startLocalServer();
-}
-
-function configureWebSocketEvent(event: APIGatewayEvent) {
-  const { requestContext } = event;
-  const { connectionId, eventType } = requestContext ?? {};
-  if (!eventType) return;
-
-  if (eventType === 'MESSAGE') {
-    const body = JSON.parse(event.body ?? '{}');
-    event.path = `/event/${body.event ?? 'unknown'}`;
-    event.httpMethod = 'POST';
-  } else {
-    event.path = `/event/${eventType.toLowerCase()}`;
-    event.httpMethod = 'GET';
-  }
-
-  (event.queryStringParameters ??= {})['connectionId'] = connectionId ?? '';
-  (event.multiValueQueryStringParameters ??= {})['connectionId'] = [connectionId ?? ''];
 }
 
 export const handler: Handler = async (event, context, callback) => {
